Fiddle Yard.c --> Changed TRISE=0xFF to 0x3F to make PORTE.bits.RE6 and 7 outputs via opto. Fiddle Yard.c --> Changed TRISD=0x00 to 0x60 to make PORTD.bits.RD5 and 6 inputs to support I^2C functionality. A test with an I^2C IO Expander will be conducted to generate extra IO. The MCP23017 will be used for this test.
Shift_Register.c --> Moved "Bezet_Weerstand" output from PORTD.bits.RD5 to PORTD.bits.RD7, to free up the I^2C Clock output.
Shift_Register.c --> At the bottom of the page the IO mapping is added to generate an overview
of used and free IO.

The code to enable and use the IO expander is not yet included.