Added visual blinking leds for TX and RX led based on TMR0. Fixed bias resistors regarding TX_ENA_LAT set in slave causing a glitch. Fixed Master behavior when a byte is received, it sets Petit_Tx_State = PETIT_RXTX_IDLE out of the PETIT_RXTX_WAIT_ANSWER state.
On 10ms interval over 4.5 hrs around 1833 messages were missed meaning missing 1 message per 880 messages.
However on 2ms update rate this is way higher. Perhaps the slaves are missing bytes due to processmodbus call is to slow?
Or the interrupt call routines via the MCC generated files are to slow! Some us and maybe you miss a byte --> to be checked and measured.