Added SPI master to TrackModbusMaster.X, per 2ms also 1 complete struct is send out over SPI.
In the final setup, the modbus master will be in init mode and not sending any data to the slaves.
During this mode the SPI slave can send all setup data for all slaves to the master.
After this the modbus slaves address need to be set via the backplabe modbus slaves.
THen the modbus master can start its routine where data is read via SPI and send back.
The last byte is the modbus member address (modbus master has address 0)

To do:
- setup SPI slave on Ethernet target
- process all data received from modbus master
- Send back data to modbus master
- try to setup startup routine